+++ packages/sheets-ui/src/services/mark-selection/mark-selection.service.ts
@@ -15,15 +15,15 @@
  */
 
 import type { Workbook } from '@univerjs/core';
+import type { ISelectionWithStyle } from '@univerjs/sheets';
 import { createIdentifier, Disposable, Inject, IUniverInstanceService, ThemeService, Tools, UniverInstanceType } from '@univerjs/core';
 import { IRenderManagerService } from '@univerjs/engine-render';
-import type { ISelectionWithStyle } from '@univerjs/sheets';
 
-// import { ISheetSelectionRenderService } from '../selection/base-selection-render.service';
-import { SelectionControl } from '../selection/selection-shape';
-import { SheetSkeletonManagerService } from '../sheet-skeleton-manager.service';
 import { ISheetSelectionRenderService } from '../selection/base-selection-render.service';
 import { SELECTION_SHAPE_DEPTH } from '../selection/const';
+// import { ISheetSelectionRenderService } from '../selection/base-selection-render.service';
+import { SelectionControl } from '../selection/selection-control';
+import { SheetSkeletonManagerService } from '../sheet-skeleton-manager.service';
 
 export interface IMarkSelectionService {
     addShape(selection: ISelectionWithStyle, exits?: string[], zIndex?: number): string | null;
@@ -57,6 +57,7 @@ export class MarkSelectionService extends Disposable implements IMarkSelectionSe
         @Inject(ThemeService) private readonly _themeService: ThemeService
     ) {
         super();
+        window.msrs = this;
     }
 
     addShape(selection: ISelectionWithStyle, exits: string[] = [], zIndex: number = DEFAULT_Z_INDEX): string | null {
