+++ packages/sheets-ui/src/services/selection/mobile-selection-shape.ts
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-import { RANGE_TYPE } from '@univerjs/core';
-import { Rect, SHEET_VIEWPORT_KEY } from '@univerjs/engine-render';
 import type { ThemeService } from '@univerjs/core';
 import type { BaseObject, IRectProps, Scene } from '@univerjs/engine-render';
 import type { ISelectionStyle } from '@univerjs/sheets';
+import { RANGE_TYPE } from '@univerjs/core';
+import { Rect, SHEET_VIEWPORT_KEY } from '@univerjs/engine-render';
 
 import { SHEET_COMPONENT_SELECTION_LAYER_INDEX } from '../../common/keys';
-import { SELECTION_MANAGER_KEY, SelectionControl } from './selection-shape';
+import { SELECTION_MANAGER_KEY, SelectionControl } from './selection-control';
 
 export class MobileSelectionControl extends SelectionControl {
     /**
