+++ packages/sheets-ui/src/services/selection/selection-shape-extension.ts
@@ -22,10 +22,10 @@ import type { IFreeze, Injector, IRangeWithCoord, Nullable, ThemeService } from 
 import type { IMouseEvent, IPointerEvent, Scene, SpreadsheetSkeleton, Viewport } from '@univerjs/engine-render';
 
 import type { Subscription } from 'rxjs';
+import type { SelectionControl } from './selection-control';
 import { SheetSkeletonManagerService } from '../sheet-skeleton-manager.service';
 import { ISheetSelectionRenderService } from './base-selection-render.service';
 import { RANGE_FILL_PERMISSION_CHECK, RANGE_MOVE_PERMISSION_CHECK } from './const';
-import type { SelectionControl } from './selection-shape';
 
 const HELPER_SELECTION_TEMP_NAME = '__SpreadsheetHelperSelectionTempRect';
